+++ src/main/java/eureca/capstone/project/orchestrator/alarm/entity/AlarmType.java
@@ -1,6 +1,6 @@
 package eureca.capstone.project.orchestrator.alarm.entity;
 
-import eureca.capstone.project.orchestrator.common.entiry.BaseEntity;
+import eureca.capstone.project.orchestrator.common.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
