+++ src/main/java/eureca/capstone/project/orchestrator/transaction_feed/entity/DataCoupon.java
@@ -1,7 +1,7 @@
 package eureca.capstone.project.orchestrator.transaction_feed.entity;
 
-import eureca.capstone.project.orchestrator.common.entiry.BaseEntity;
-import eureca.capstone.project.orchestrator.common.entiry.TelecomCompany;
+import eureca.capstone.project.orchestrator.common.entity.BaseEntity;
+import eureca.capstone.project.orchestrator.common.entity.TelecomCompany;
 import jakarta.persistence.*;
 
 @Entity
