+++ src/main/java/eureca/capstone/project/orchestrator/transaction_feed/entity/TransactionFeed.java
@@ -1,8 +1,8 @@
 package eureca.capstone.project.orchestrator.transaction_feed.entity;
 
-import eureca.capstone.project.orchestrator.common.entiry.BaseEntity;
-import eureca.capstone.project.orchestrator.common.entiry.Status;
-import eureca.capstone.project.orchestrator.common.entiry.TelecomCompany;
+import eureca.capstone.project.orchestrator.common.entity.BaseEntity;
+import eureca.capstone.project.orchestrator.common.entity.Status;
+import eureca.capstone.project.orchestrator.common.entity.TelecomCompany;
 import eureca.capstone.project.orchestrator.user.entity.User;
 import jakarta.persistence.*;
