+++ src/main/java/eureca/capstone/project/orchestrator/user/entity/Plan.java
@@ -1,8 +1,8 @@
 package eureca.capstone.project.orchestrator.user.entity;
 
 
-import eureca.capstone.project.orchestrator.common.entiry.BaseEntity;
-import eureca.capstone.project.orchestrator.common.entiry.TelecomCompany;
+import eureca.capstone.project.orchestrator.common.entity.BaseEntity;
+import eureca.capstone.project.orchestrator.common.entity.TelecomCompany;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
