+++ src/main/java/eureca/capstone/project/orchestrator/user/entity/UserData.java
@@ -1,6 +1,6 @@
 package eureca.capstone.project.orchestrator.user.entity;
 
-import eureca.capstone.project.orchestrator.common.entiry.BaseEntity;
+import eureca.capstone.project.orchestrator.common.entity.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
